requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.7

# Microsoft Graph API
msal==1.22.0
requests==2.30.0

# MongoDB
pymongo==4.3.3

# Scheduling
apscheduler==3.10.1

# Utilities
python-multipart==0.0.6
email-validator==2.0.0